AI Project copy/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
python-dotenv==1.0.1
requests==2.31.0
apscheduler==3.10.4
groq==0.5.0
pytest==8.1.1
httpx==0.27.0
